Move config models to internal
