+++ hw12_13_14_15_calendar/internal/config/calendar.go
@@ -1,0 +1,19 @@
+package config
+
+type Config struct {
+	Logger LoggerConf
+	// TODO
+}
+
+type LoggerConf struct {
+	Level string
+	// TODO
+}
+
+func NewConfig(logLvl string) *Config {
+	return &Config{
+		Logger: LoggerConf{
+			Level: logLvl,
+		},
+	}
+}
